requirements.txt
Flask==1.0.2
gunicorn==19.9.0  
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
numpy==1.16.2
Pillow==5.4.1
PyYAML==5.1
tensorflow==1.13.1
